Concurrency fix on the parser. Version bump.
Now the Parser correcly handles multiple requests of parsing of the same location from multiple
threads. Just one sub-thread parses the location and the answer is sent back on a multiple tapped
channel (the thread which parses writes on a clojure.core.async/mult).  The tests have been adjsted
accordingly and they are all green.